Enable alpha blending for ID pass (https://github.com/enso-org/ide/pull/364)
* Enable working alpha blending for ID pass via bit encoding of IDs in a RGB8 texture..
* Add facilities to change texture parameters.

Original commit: https://github.com/enso-org/ide/commit/b46012e3fa57b674ff03c8c360d24b2e544f0590

+++ ide/src/rust/ensogl/src/display/shape/text/glyph/system.rs
@@ -258,7 +258,7 @@ impl GlyphSystem {
         // FIXME outputs as the number of attachments to framebuffer. We should manage this more
         // FIXME intelligent. For example, we could allow defining output shader fragments,
         // FIXME which will be enabled only if pass of given attachment type was enabled.
-        material.add_output("id", Vector4::<u32>::new(0,0,0,0));
+        material.add_output("id", Vector4::<f32>::new(0.0,0.0,0.0,0.0));
 
         let code = CodeTemplate::new(BEFORE_MAIN.to_string(),MAIN.to_string(),"".to_string());
         material.set_code(code);
